refactor: test case for eod sync and manufacturing operation
