+++ .github/workflows/test.yml
@@ -159,13 +159,6 @@ jobs:
           bench --site test_site run-tests \
             --module jewellery_erpnext.jewellery_erpnext.tests.test_make_receive_entry_mop_cap
 
-      - name: Run Manufacturing Operation Weight Calculation 
-        run: |
-          cd ~/frappe-bench
-
-          bench --site test_site run-tests \
-            --module jewellery_erpnext.jewellery_erpnext.tests.test_mop_weight_recompute
-
       - name: Run Department IR
         run: |
           cd ~/frappe-bench
@@ -193,13 +186,6 @@ jobs:
 
           bench --site test_site run-tests \
             --app jewellery_erpnext --doctype "MOP Log"
-
-      - name: Run EOD Sync
-        run: |
-          cd ~/frappe-bench
-
-          bench --site test_site run-tests \
-            --module jewellery_erpnext.jewellery_erpnext.tests.test_eod_sync_gap_fill
 
       - name: Run MOP Weight Calculation
         run: |
